style: fix minor lint issues (RUF046, G201, RUF059, W292)
- backend/core/rate_limit.py: remove redundant int() cast on math.ceil
- backend/scheduler/__init__.py: use .exception() instead of .error(exc_info)
- backend/utils/tg_session.py: prefix unused api_id with underscore
- tg_signer/core.py: prefix unused protocol with underscore
- tests/test_config_import.py: add missing trailing newline

+++ backend/core/rate_limit.py
@@ -45,7 +45,7 @@ class InMemoryRateLimiter:
         with self._lock:
             blocked_until = self._blocked_until.get(bucket, 0.0)
             if blocked_until > now:
-                retry_after = max(int(math.ceil(blocked_until - now)), 1)
+                retry_after = max(math.ceil(blocked_until - now), 1)
                 raise HTTPException(
                     status_code=status.HTTP_429_TOO_MANY_REQUESTS,
                     detail=detail,
